// services/core/types.ts
export interface CartItem {
  name: string;
  quantity: number;
  price: number;
  category: string;
  itemIndex: number;
}

// Extender BaseConversation para incluir los campos específicos del carrito
export interface CartConversation extends BaseConversation {
  cart: CartItem[];
  selectedCategory?: string;
  selectedItem?: string;
  selectedItemIndex?: number;
}

interface QuantitySelectionProps {
  conversation: CartConversation;
  quantity: number;
  price: number;
  deliveryCost: number;
  updateConversationFn: (updatedCart: CartItem[]) => Promise<void>;
}

export type QuantitySelectionFn = (Props: QuantitySelectionProps) => Promise<string>;

interface CartActionsProps {
  conversation: CartConversation;
  option: number;
  deliveryCost: number;
  transfersPhoneNumber: string;
  updateConversationFn: (updates: Partial<CartConversation>) => Promise<void>;
  welcomeMessageFn: GetWelcomeMessageFn;
  addMoreItemsFn: () => string;
}

export type CartActionsFn = (Props: CartActionsProps) => Promise<string>;

interface CheckoutMessageProps {
  cart: CartItem[];
  deliveryCost: number;
  transfersPhoneNumber: string;
}

export type CheckoutMessageFn = (Props: CheckoutMessageProps) => string;

interface CheckoutProps {
  conversation: CartConversation;
  option: number;
  deliveryCost: number;
  transfersPhoneNumber: string;
  updateConversationFn: (updates: Partial<CartConversation>) => Promise<void>;
  welcomeMessageFn: GetWelcomeMessageFn;
  addMoreItemsFn: () => string;
  finalMessageFn: () => Promise<string>;
}

export type CheckoutFn = (Props: CheckoutProps) => Promise<string>;
// ...
export type GetWelcomeMessageFn = (msgPreliminar?: string, greeting?: boolean) => string;